+++ packages/react-theme/src/themes/teams/darkTheme.ts
@@ -1,5 +1,5 @@
-import { createNeutralColorTokens, sharedColorTokens } from '../../dark/index';
-import { createShadowLevelTokens } from '../../default/index';
+import { createNeutralColorTokens, sharedColorTokens } from '../../utils/dark/index';
+import { createShadowLevelTokens } from '../../utils/light/index';
 import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights, grey } from '../../global/index';
 import { Theme } from '../../types';
