+++ packages/react-theme/src/themes/teams/lightTheme.ts
@@ -1,4 +1,4 @@
-import { createNeutralColorTokens, sharedColorTokens, createShadowLevelTokens } from '../../default/index';
+import { createNeutralColorTokens, sharedColorTokens, createShadowLevelTokens } from '../../utils/light/index';
 import { brandColors, sharedColors, fontFamilies, fontWeights, fontSizes, lineHeights } from '../../global/index';
 import { Theme } from '../../types';
 
@@ -13,7 +13,7 @@ const common = {
 
 const teamsNeutralColorTokens = createNeutralColorTokens(brandColors.teams);
 
-export const teamsDefaultTheme: Theme = {
+export const teamsLightTheme: Theme = {
   brandColors: brandColors.teams,
   neutralColorTokens: teamsNeutralColorTokens,
   shadowLevels: createShadowLevelTokens(
