+++ packages/react-theme/src/themes/web/index.ts
@@ -1,3 +1,3 @@
-export * from './defaultTheme';
+export * from './lightTheme';
 export * from './darkTheme';
 export * from './highContrastTheme';
