+++ packages/react-theme/src/utils/highContrast/index.ts
